Moved update js db script to updatedb hook implementation for js_file plugin. Deleted original script.
