--- scripts/js_libraries/get_retirejs_db.py
@@ -1,41 +1,0 @@
-import json
-
-import requests
-
-
-def load_database() -> dict:
-    url = "https://raw.githubusercontent.com/RetireJS/retire.js/master/repository/jsrepository-master.json"
-    response = requests.get(url)
-    if response.status_code == 200:
-        return json.loads(response.text)
-    return None
-
-
-def strip_irrelevant_data(retirejs_db: dict) -> dict:
-    clean_db = {}
-    reg_temp = "\u00a7\u00a7version\u00a7\u00a7"
-    version_regex = r"\d+(?:\.\d+)*"
-    for library, lib_entry in retirejs_db.items():
-        if "extractors" in lib_entry:
-            clean_db[library] = {}
-            patterns = lib_entry["extractors"]
-            possible_entries = [
-                "filename",
-                "filecontent",
-                "hashes",
-            ]
-            for entry in possible_entries:
-                if entry in patterns:
-                    entry_list = []
-                    for reg in patterns[entry]:
-                        entry_list.append(reg.replace(reg_temp, version_regex))
-                    clean_db[library][entry] = entry_list
-    return clean_db
-
-
-retirejs = load_database()
-
-if retirejs is not None:
-    cleaned = strip_irrelevant_data(retirejs)
-    with open("js_library_patterns.json", "w") as f:
-        json.dump(cleaned, f, indent=4)
