+++ surfactant/infoextractors/js_file.py
@@ -11,7 +11,7 @@ from loguru import logger
 
 import surfactant.plugin
 from surfactant.sbomtypes import SBOM, Software
-
+import requests
 
 def supports_file(filetype) -> bool:
     return filetype == "JAVASCRIPT"
@@ -67,6 +67,44 @@ def match_by_attribute(attribute: str, content: str, database: Dict) -> List[Dic
                         # skip remaining patterns, move on to the next library
                         break
     return libs
+
+def load_database() -> dict:
+    url = "https://raw.githubusercontent.com/RetireJS/retire.js/master/repository/jsrepository-master.json"
+    response = requests.get(url)
+    if response.status_code == 200:
+        return json.loads(response.text)
+    return None
+
+
+def strip_irrelevant_data(retirejs_db: dict) -> dict:
+    clean_db = {}
+    reg_temp = "\u00a7\u00a7version\u00a7\u00a7"
+    version_regex = r"\d+(?:\.\d+)*"
+    for library, lib_entry in retirejs_db.items():
+        if "extractors" in lib_entry:
+            clean_db[library] = {}
+            patterns = lib_entry["extractors"]
+            possible_entries = [
+                "filename",
+                "filecontent",
+                "hashes",
+            ]
+            for entry in possible_entries:
+                if entry in patterns:
+                    entry_list = []
+                    for reg in patterns[entry]:
+                        entry_list.append(reg.replace(reg_temp, version_regex))
+                    clean_db[library][entry] = entry_list
+    return clean_db
+
+
+@surfactant.plugin.hookimpl
+def update_db():
+    retirejs = load_database()
+    if retirejs is not None:
+        cleaned = strip_irrelevant_data(retirejs)
+        with open("js_library_patterns.json", "w") as f:
+            json.dump(cleaned, f, indent=4)
 
 
 @surfactant.plugin.hookimpl
